pyproject.toml
[tool.poetry]
name = "advanced-rag"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "<2"
python-dotenv = "^1.0.1"
langchain-community = "^0.0.24"
tiktoken = "^0.6.0"
chromadb = "^0.4.23"
langchain = "^0.1.9"
bs4 = "^0.0.2"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
